refactor: move editorItemCallback to anonymous #85
